+++ src-tauri/Cargo.toml
@@ -4,7 +4,7 @@ name = "vault"
 version = "0.1.0"
 description = "A Tauri App"
 authors = ["you"]
-edition = "2024"
+edition = "2021"
 
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
